+++ addon/components/item-picker/component.js
@@ -2,29 +2,18 @@ import Ember from 'ember';
 import layout from './template';
 import queryHelpers from 'ember-arcgis-portal-components/utils/query-helpers';
 import isGuid from 'ember-arcgis-portal-components/utils/is-guid';
-
 export default Ember.Component.extend({
-
   layout,
-
   intl: Ember.inject.service(),
-
   itemService: Ember.inject.service('items-service'),
-
-  _i18nScope: Ember.computed('i18nScope', function () {
-    return `${this.getWithDefault('i18nScope', 'ember-arcgis-portal-components.itemPicker')}.`;
-  }),
-
   classNames: [ 'item-picker', 'clearfix', 'row' ],
 
-  isValidating: false,
-
-  shouldValidate: false,
-
-  selectAnyway: false,
-
+  /**
+   * Startup the component... we may need to issue an immediate search...
+   */
   init () {
     this._super(...arguments);
+    this.set('itemsToAdd', []);
     if (this.get('searchItemsOnInit')) {
       if (this.get('catalog')) {
         this._setInitialCatalog(this.get('catalog'));
@@ -33,49 +22,46 @@ export default Ember.Component.extend({
     }
   },
 
-  selectButtonText: Ember.computed('isValidating', 'selectAnyway', function () {
-    const intl = this.get('intl');
-    if (this.get('isValidating')) {
-      return intl.t(`${this.get('_i18nScope')}buttons.validating`);
-    } else if (this.get('selectAnyway')) {
-      return intl.t(`${this.get('_i18nScope')}buttons.selectAnyway`);
-    }
-    return intl.t(`${this.get('_i18nScope')}buttons.select`);
+  disableAddItems: Ember.computed.not('hasItemsToAdd'),
+  showNoItemsMsg: Ember.computed.notEmpty('noItemsFoundMsg'),
+  hasItemsToAdd: Ember.computed.notEmpty('itemsToAdd'),
+  isValidating: false,
+  selectAnyway: false,
+  shouldValidate: false,
+
+  /**
+   * Compute the translation scope
+   */
+  _i18nScope: Ember.computed('i18nScope', function () {
+    return `${this.getWithDefault('i18nScope', 'ember-arcgis-portal-components.itemPicker')}.`;
   }),
 
-  selectButtonClass: Ember.computed('isValidating', 'selectAnyway', 'errorMessage', function () {
-    const errorMessage = this.get('errorMessage');
-    if (this.get('isValidating')) {
-      return 'disabled';
-    } else if (errorMessage && errorMessage.status && errorMessage.status === 'error') {
-      return 'disabled';
+  /**
+   * Allow a loading component to be specified as a parameter
+   */
+  loadingComponent: Ember.computed('loadingComponentName', function () {
+    let result = 'loading-indicator';
+    if (this.get('loadingComponentName')) {
+      result = this.get('loadingComponentName');
     }
+    return result;
   }),
 
-  previewUrl: Ember.computed('portalHostName', 'currentItem.id', 'currentItem.url', function () {
-    const portalHostName = this.get('portalHostName');
-    const id = this.get('currentItem.id');
-    const type = this.get('currentItem.type');
-    const url = this.get('currentItem.url');
-    const protocol = '//';
-    let host = (portalHostName === undefined) ? this.get('session.portalHostname') : portalHostName;
-    let previewURL;
+  /**
+   * Determine what preview component to use. This allows us to create
+   * per-type UX for the preview
+   */
+  previewComponent: Ember.computed('currentItem', function () {
+    let type = this.get('currentItem.type');
+    let componentName = 'item-picker/item-preview';
 
-    switch (true) {
-      // Is there a URL? If so, use it
-      case url !== null:
-        previewURL = this.get('currentItem.url');
+    switch (type.toLowerCase()) {
+      case 'feature service':
+      case 'map service':
+        componentName = 'item-picker/feature-service-preview';
         break;
-      // Is it a webmap? If so, send a webmap specific link
-      case type.toLowerCase() === 'web map':
-        previewURL = `${protocol}${host}/home/webmap/viewer.html?webmap=${id}`;
-        break;
-      // Otherwise, just return the item page
-      default:
-        previewURL = `${protocol}${host}/home/item.html?id=${id}`;
     }
-
-    return previewURL;
+    return componentName;
   }),
 
   inputElementId: Ember.computed(function () {
@@ -104,7 +90,10 @@ export default Ember.Component.extend({
 
   hasSearched: false,
 
-  showTabs: Ember.computed('catalog', function () {
+  /**
+   * Do we show facets? if we have more than one entry in the catalog, yes
+   */
+  showFacets: Ember.computed('catalog', function () {
     const catalog = this.get('catalog');
     if (catalog && catalog.length > 1) {
       return true;
@@ -142,14 +131,6 @@ export default Ember.Component.extend({
     }
   }),
 
-  showNoItemsMsg: Ember.computed.notEmpty('noItemsFoundMsg'),
-
-  itemsToAdd: [],
-
-  hasItemsToAdd: Ember.computed.notEmpty('itemsToAdd'),
-
-  disableAddItems: Ember.computed.not('hasItemsToAdd'),
-
   _defaultSearch (q, isValidGuid) {
     let parts = [];
     if (q) {
@@ -171,6 +152,9 @@ export default Ember.Component.extend({
     return parts.join(' AND ');
   },
 
+  /**
+   * Execute the search
+   */
   _doSearch (q, page = 1) {
     let isValidGuid = isGuid(q);
     this.setProperties({
@@ -209,6 +193,10 @@ export default Ember.Component.extend({
       });
   },
 
+  /**
+   * Developer can pass in a .active property on a catalog entry,
+   * this code simply ensures that is the "active" catalog
+   */
   _setInitialCatalog (catalog) {
     let startingCatalog = catalog[0];
     let selectedCatalog = catalog.reduce((acc, entry) => {
@@ -223,6 +211,9 @@ export default Ember.Component.extend({
   },
 
   actions: {
+    /**
+     * Fired when a facet is selected
+     */
     chooseCatalog (val) {
       const selectedCatalog = this.get('catalog').findBy('name', val);
       this.set('selectedCatalogName', selectedCatalog.name);
@@ -231,16 +222,25 @@ export default Ember.Component.extend({
       this._doSearch(this.get('q'));
     },
 
+    /**
+     * When the user types in the search box, this fires...
+     */
     doSearch (query) {
       const q = query;
       this.set('q', q);
       Ember.run.debounce(this, this._doSearch, q, 150);
     },
+    /**
+     * Paging
+     */
     changePage (page) {
       const q = this.get('q');
       this._doSearch(q, page);
     },
-    onClick (item) {
+    /**
+     * When an item is clicked in the list
+     */
+    onItemClick (item) {
       if (this.get('selectMultiple')) {
         const itemsToAdd = this.get('itemsToAdd');
         const existingObj = itemsToAdd.findBy('id', item.id);
@@ -261,29 +261,12 @@ export default Ember.Component.extend({
         }
       }
     },
-
-    onSelect (item) {
-      const validator = this.get('onSelectionValidator');
-
-      this.set('isValidating', true);
-      if (validator && typeof validator === 'function' && !this.get('selectAnyway')) {
-        validator(item)
-          .then((resp) => {
-            this.set('isValidating', false);
-            this.set('errorHash', resp.status);
-            if (resp.status.status === 'error') {
-              return;
-            } else if (resp.status.status === 'warning') {
-              this.set('selectAnyway', true);
-              return;
-            } else {
-              return this.get('selectAction')(resp.item);
-            }
-          });
-      } else {
-        this.set('isValidating', false);
-        return this.get('selectAction')(item);
-      }
+    /**
+     * Called when user clicks "Select" in the Preview component
+     */
+    onPreviewSelected (item, options) {
+      // call the closure action passed into this component
+      return this.get('selectAction')(item, options);
     },
 
     cancelAction () {