+++ app/components/item-picker/layer-picker/component.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-arcgis-portal-components/components/item-picker/layer-picker/component';